pyproject.toml
[project]
name = "meister-eder"
version = "0.1.0"
description = "AI-powered conversational registration agent for Spielgruppe Pumuckl"
requires-python = ">=3.13"
dependencies = [
    # LLM access — supports any provider (Anthropic, OpenAI, Gemini, …)
    "litellm>=1.0.0",
    # Configuration
    "python-dotenv>=1.0.0",
    # Registration schema validation
    "jsonschema>=4.23.0",
    "chainlit>=2.9.6",
    "qrbill>=1.2.0",
    "pillow>=12.1.1",
    "jinja2>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
meister-eder = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
